requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
channels==2.4.0
channels-redis==2.4.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==2.5.0
Deprecated==1.2.13
Django==4.1.1
django-ckeditor==6.5.1
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-quill-editor==0.1.40
djangorestframework==3.13.1
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.0
facebook-sdk==3.1.0
Faker==18.7.0
firebase-admin==6.0.1
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==0.6.2
packaging==21.3
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
xlwt==1.3.0
zope.interface==5.5.1
